add downloads volume backup
adding this for posterity, but no longer required. This was originally used to backup the downloads volume which was a separate rook-ceph volume because the windows ntfs storage couldn't support links. Now that we've migrated fully to linux and the nfs, the media volume can accept soft (and hard) links, so the downloads volume can now be deleted.